Bugfix/dpi independent html alignment

+++ 3DAmsterdam/Assets/Netherlands3D/Scripts/JavascriptInteraction/JavascriptMethodCaller.cs
@@ -6,7 +6,7 @@ namespace Netherlands3D.JavascriptConnection
 	public class JavascriptMethodCaller
 	{
 		[DllImport("__Internal")]
-		private static extern void DisplayDOMObjectWithID(string id = "htmlID", string display = "none", int x = 0, int y = 0, int width = 0, int height = 0);
+		private static extern void DisplayDOMObjectWithID(string id = "htmlID", string display = "none", float x = 0, float y = 0, float width = 0, float height = 0);
 
 		[DllImport("__Internal")]
 		private static extern string FetchOBJData();
@@ -49,7 +49,7 @@ namespace Netherlands3D.JavascriptConnection
 		/// Faking a click through JavaScript is not allowed.
 		/// </summary>
 		/// <param name="display">Sets the hitarea CSS of the input HTML node to inline, or none</param>
-		public static void DisplayWithID(string id, bool display, int x = 0, int y = 0, int width = 0, int height = 0 )
+		public static void DisplayWithID(string id, bool display, float x = 0, float y = 0, float width = 0, float height = 0 )
 		{
 #if UNITY_WEBGL && !UNITY_EDITOR
 			 DisplayDOMObjectWithID(id,(display) ? "inline" : "none", x, y, width, height);
